requirements.txt
discord.py==2.3.2
aiohttp==3.9.4
PyNaCl==1.5.0
aiosqlite==0.19.0
ffmpeg==1.4